pyproject.toml
[tool.poetry]
name = "ecli"
version = "0.1.0"
description = ""
authors = ["Pieterjan Montens <pieterjan@montens.net>"]
packages = [
    { include = "webservice" },
]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.1"
PyYAML = "^5.3.1"
pytz = "^2020.4"
asyncpg = "^0.21.0"
uvicorn = "^0.12.2"
msgpack = "^1.0.0"
requests = "^2.24.0"

[tool.poetry.dev-dependencies]
pylint = "^2.6.0"
flake8 = "^3.8.4"
isort = "^5.6.4"

[tool.poetry.scripts]
api = "webservice.main:main"

[tool.dephell.main]
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "setuppy", path = "setup.py"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
